+++ demo/cypress/DropZone/index.tsx
@@ -1,5 +1,7 @@
+import { isNil } from 'lodash';
 import * as React from 'react';
 import * as L from '../../../leda';
+import { useInterval } from '../../../leda/utils';
 
 export const DropZone = (): React.ReactElement => {
   const [value, setValue] = React.useState<L.DropZoneTypes.DropZoneFiles | null>({
@@ -7,7 +9,14 @@ export const DropZone = (): React.ReactElement => {
     rejectedFiles: [],
   });
   const [isDisabled, setIsDisabled] = React.useState<boolean>(false);
+  const [isLoading, setIsLoading] = React.useState<boolean>(false);
+  const [loadingProgress, setLoadingProgress] = React.useState<number | undefined>(undefined);
+  const [isCustomLoader, setIsCustomLoader] = React.useState<boolean>(false);
 
+  useInterval(() => {
+    setLoadingProgress((progress) => progress === 100 ? 0 : progress + 5);
+  }, !isNil(loadingProgress) ? 100 : null);
+  
   return (
     <L.Div _demoStory>
       <L.Span>Uncontrolled, without validation</L.Span>
@@ -28,14 +37,27 @@ export const DropZone = (): React.ReactElement => {
       <L.Span>Controlled</L.Span>
       <L.DropZone
         value={value}
+        isLoading={isLoading}
+        loadingProgress={loadingProgress}
+        loadingViewRender={isCustomLoader && (() => <L.Span _txtSuccess>Custom loader...</L.Span>)}
         onChange={({ component: { value } }) => setValue(value)}
         isRequired
         isDisabled={isDisabled}
       />
       <br />
-      <L.Button onClick={() => setValue(null)}>Set state as null</L.Button>
+      <L.Button _warning onClick={() => setValue(null)}>Set state as null</L.Button>
       <br />
-      <L.Button name="disable" onClick={() => setIsDisabled(!isDisabled)}>Disable / Enable</L.Button>
+      <br />
+      <L.Button _warning name="disable" onClick={() => setIsDisabled(!isDisabled)}>Disable / Enable</L.Button>
+      <br />
+      <br />
+      <L.Button _warning name="loader" onClick={() => setIsLoading(!isLoading)}>Show/hide loader</L.Button>
+      <br />
+      <br />
+      <L.Button _warning name="progressLoader" onClick={() => {setLoadingProgress(progress => isNil(progress) ? 0 : undefined); setIsLoading(!isLoading); }}>Show/hide progress</L.Button>
+      <br />
+      <br />
+      <L.Button _warning name="customLoader" onClick={() => {setIsCustomLoader(!isCustomLoader); setIsLoading(!isLoading); }}>Set/remove Custom Loader</L.Button>
       <br />
 
       <L.Span>Uncontrolled (prop value is null)</L.Span>