+++ leda/components/DropZone/theme.ts
@@ -9,9 +9,13 @@ export const defaultDropZoneTheme = {
   fileDeleteLink:       'dropzone-file-delete internal',
   fileDownloadLink:     'dropzone-file-download internal',
   invalid:              'danger',
+  loader:               'dropzone-loader',
+  progressBase:         'progress-base',
+
+  progressCircle:       'progress-circle',
+  progressLoader:       'dropzone-progress-loader',
   rejectedFilesWrapper: 'rejected-list margin-none',
   rejectedMessage:      'dropzone-rejected-message txt-danger',
-
   required:             'required',
   status:               'dropzone-status',
   wrapper:              'dropzone-wrapper',