+++ CHANGELOG.md
@@ -9,6 +9,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Fixed 
 - Validation: Fixed displaying of invalidMessage with isValid prop set to false
 
+### Added
+- DropZone: Added isLoading, loadingProgress and loadingViewRender props
+
 ## [0.34.0] - 2020-11-09
 
 ### Added